+++ terms_query_result.go
@@ -3,6 +3,6 @@ package ravendb
 // TermsQueryResult represents results of terms query
 type TermsQueryResult struct {
 	Terms      []string `json:"Terms"`
-	ResultEtag int      `json:"ResultEtag"`
+	ResultEtag int64    `json:"ResultEtag"`
 	IndexName  string   `json:"IndexName"`
 }
