+++ topology.go
@@ -3,5 +3,5 @@ package ravendb
 // Topology describes server nodes
 type Topology struct {
 	Nodes []*ServerNode `json:"Nodes"`
-	Etag  int           `json:"Etag"`
+	Etag  int64         `json:"Etag"`
 }
